Create container and reducer
