+++ src/components/ItemRecognition/actionTypes.js
