+++ src/components/ItemRecognition/reducer.js
@@ -1,4 +1,4 @@
-import {SET_PREDICTION} from './actionTypes';
+import {SET_PREDICTION} from '../ItemRecognition/actionTypes';
 
 export function prediction(state = null, action) {
 	if (action.type === SET_PREDICTION) return action.prediction;
