--- src/components/WebcamCapture/WebcamCaptureContainer.js
@@ -1,20 +1,0 @@
-import {connect} from 'react-redux';
-import WebcamCapture from './WebcamCapture';
-import {setPrediction} from './actions';
-
-const mapStateToProps = state => {
-	return {prediction: state.prediction};
-};
-
-const mapDispatchToProps = dispatch => {
-	return {
-		setPrediction: (id, img) => {
-			dispatch(setPrediction({id, img}));
-		}
-	};
-};
-
-export default connect(
-	mapStateToProps,
-	mapDispatchToProps
-)(WebcamCapture);
